Valiu_User_Inyerface_Challenge/requirements.txt
behave==1.2.6
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
idna==2.10
parse==1.18.0
parse-type==0.5.2
PyAutoIt==0.6.0
requests==2.25.0
selenium==3.141.0
six==1.15.0
urllib3==1.26.2
webdriver-manager==3.2.2
